chore: add eslint and prettier
