+++ .eslintrc
@@ -1,0 +1,15 @@
+{
+  "extends": ["react-app", "react-app/jest", "airbnb", "airbnb/hooks", "prettier"],
+  "plugins": ["prettier"],
+  "ignorePatterns": ["node_modules", "src/serviceWorker.js"],
+  "rules": {
+    "func-names": "off",
+    "no-console": "warn",
+    "react/prop-types": "off",
+    "prettier/prettier": "error",
+    "react/react-in-jsx-scope": "off",
+    "import/prefer-default-export": "off",
+    "jsx-a11y/label-has-associated-control":"off",
+    "jsx-a11y/click-events-have-key-events": "off"
+  }
+}
