--- src/Components/Logo.js
@@ -1,10 +1,0 @@
-function Logo({ img, title, alt }) {
-  return (
-    <a className="logo" href="/">
-      <img src={img} alt={alt} />
-      <p>{title}</p>
-    </a>
-  );
-}
-
-export default Logo;
