+++ src/components/Logo.jsx
@@ -1,0 +1,10 @@
+function Logo({ img, title, alt }) {
+  return (
+    <a className="logo" href="/">
+      <img src={img} alt={alt} />
+      <p>{title}</p>
+    </a>
+  );
+}
+
+export default Logo;
