+++ src/components/MessageList.jsx
@@ -10,8 +10,8 @@ function MessageList({ stop, phoneNumbers, message }) {
   };
 
   const sendMessage = () => {
-    let phoneNumber = phoneNumbers[activeIndex];
-    let newNumber = phoneNumber.toString().length === 10 ? `+91${phoneNumber}` : phoneNumber;
+    const phoneNumber = phoneNumbers[activeIndex];
+    const newNumber = phoneNumber.toString().length === 10 ? `+91${phoneNumber}` : phoneNumber;
 
     const link = `https://api.whatsapp.com/send?phone=${newNumber}${
       message && `&text=${encodeURIComponent(message)}`
