+++ src/components/PersonalMessage.jsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
 import PhoneInput from 'react-phone-input-2';
+
 import MessageBox from './MessageBox';
 import 'react-phone-input-2/lib/bootstrap.css';
 
@@ -7,32 +8,30 @@ function MessageForm() {
   const [phoneNumber, setPhoneNumber] = useState('');
   const [message, setMessage] = useState('');
   const [error, setError] = useState('');
-  let link = `https://api.whatsapp.com/send?phone=${phoneNumber}${
+  const link = `https://api.whatsapp.com/send?phone=${phoneNumber}${
     message && `&text=${encodeURIComponent(message)}`
   }`;
 
-  let handleLinkClick = () => {
-    if (validatePhoneNumber() && validateMessage()) {
-      window.location.assign(link);
-    }
-  };
-
-  let validatePhoneNumber = () => {
+  const validatePhoneNumber = () => {
     // eslint-disable-next-line
     if (phoneNumber.match(/^[\+]?[(]?[0-9]{3}[)]?[-\s\.]?[0-9]{3}[-\s\.]?[0-9]{4,6}$/)) {
       return true;
-    } else {
-      setError('Invalid Phone Number');
-      return false;
     }
+    setError('Invalid Phone Number');
+    return false;
   };
 
-  let validateMessage = () => {
+  const validateMessage = () => {
     if (message.length < 250) {
       return true;
-    } else {
-      setError('Message can contain only upto 250 characters');
-      return false;
+    }
+    setError('Message can contain only upto 250 characters');
+    return false;
+  };
+
+  const handleLinkClick = () => {
+    if (validatePhoneNumber() && validateMessage()) {
+      window.location.assign(link);
     }
   };
 
@@ -43,7 +42,8 @@ function MessageForm() {
       <label>
         Phone Number
         <PhoneInput
-          country={'in'}
+          name="phone_input"
+          country="in"
           value={phoneNumber}
           placeholder="+91 987654321"
           onChange={(phone) => setPhoneNumber(phone)}
@@ -53,10 +53,10 @@ function MessageForm() {
         name="Message"
         placeholder="Message (Optional)"
         value={message}
-        onChange={(message) => setMessage(message)}
+        onChange={(_message) => setMessage(_message)}
       />
 
-      <button onClick={handleLinkClick} className="message-btn">
+      <button type="submit" onClick={handleLinkClick} className="message-btn">
         Send Message
       </button>
     </div>
