--- src/constants.js
@@ -1,4 +1,0 @@
-let appName = 'WhatSend';
-let gitRepo = 'https://github.com/aromalanil/whatsend';
-
-export { appName, gitRepo };
