+++ src/utils/constants.js
@@ -1,0 +1,2 @@
+export const appName = 'WhatSend';
+export const gitRepo = 'https://github.com/aromalanil/whatsend';
